# perf(matrices): use DomainMatrix for inv_mod