--- a/sympy/matrices/repmatrix.py
+++ b/sympy/matrices/repmatrix.py
@@ -7,14 +7,15 @@ from sympy.core.kind import Kind, NumberKind, UndefinedKind
 from sympy.core.numbers import Integer, Rational
 from sympy.core.sympify import _sympify, SympifyError
 from sympy.core.singleton import S
-from sympy.polys.domains import ZZ, QQ, EXRAW
+from sympy.polys.domains import ZZ, QQ, GF, EXRAW
 from sympy.polys.matrices import DomainMatrix
+from sympy.polys.matrices.exceptions import DMNonInvertibleMatrixError
 from sympy.polys.polyerrors import CoercionFailed
 from sympy.utilities.exceptions import sympy_deprecation_warning
 from sympy.utilities.iterables import is_sequence
-from sympy.utilities.misc import filldedent
+from sympy.utilities.misc import filldedent, as_int
 
-from .common import classof
+from .common import classof, NonSquareMatrixError, NonInvertibleMatrixError
 from .matrices import MatrixBase, MatrixKind, ShapeError
 
 
@@ -392,6 +393,49 @@ class RepMatrix(MatrixBase):
                 elif ans is not True and rv is True:
                     rv = ans
         return rv
+
+    def inv_mod(M, m):
+        r"""
+        Returns the inverse of the integer matrix ``M`` modulo ``m``.
+
+        Examples
+        ========
+
+        >>> from sympy import Matrix
+        >>> A = Matrix(2, 2, [1, 2, 3, 4])
+        >>> A.inv_mod(5)
+        Matrix([
+        [3, 1],
+        [4, 2]])
+        >>> A.inv_mod(3)
+        Matrix([
+        [1, 1],
+        [0, 1]])
+
+        """
+
+        if not M.is_square:
+            raise NonSquareMatrixError()
+
+        try:
+            m = as_int(m)
+        except ValueError:
+            raise TypeError("inv_mod: modulus m must be an integer")
+
+        K = GF(m, symmetric=False)
+
+        try:
+            dM = M.to_DM(K)
+        except CoercionFailed:
+            raise ValueError("inv_mod: matrix entries must be integers")
+
+        try:
+            dMi = dM.inv()
+        except DMNonInvertibleMatrixError as exc:
+            msg = f'Matrix is not invertible (mod {m})'
+            raise NonInvertibleMatrixError(msg) from exc
+
+        return dMi.to_Matrix()
 
     def lll(self, delta=0.75):
         """LLL-reduced basis for the rowspace of a matrix of integers.