<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->


#### Brief description of what is fixed or changed

Use DomainMatrix over GF(p) for the Matrix.inv_mod method. This makes inv_mod significantly faster:
```
In [1]: M = randMatrix(20)

In [2]: %time ok = M.inv_mod(163)
CPU times: user 70.8 ms, sys: 0 ns, total: 70.8 ms
Wall time: 69.3 ms
```
That takes 30 seconds with master so this is 500x faster but the difference increases for larger matrices because the algorithm used on master is like `O(2^n)` rather than `O(n^3)` for DomainMatrix.

#### Other comments

I'm a little unsure really where these matrix methods belong any more. I've moved this method to RepMatrix because it uses DomainMatrix and RepMatrix is the Matrix class that is based on DomainMatrix.

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers. Formerly, `log(-x)` incorrectly gave `-log(x)`.

* physics.units
  * Corrected a semantical error in the conversion between volt and statvolt which
    reported the volt as being larger than the statvolt.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* matrices
    * The Matrix.inv_mod method was made significantly faster by using DomainMatrix and the GF(p) domain.
<!-- END RELEASE NOTES -->