@oscarbenjamin 2023-11-13
Using python-flint makes this faster e.g. with ground types set to flint:
```python
In [6]: M = randMatrix(100)

In [7]: %time ok = M.inv_mod(163)
CPU times: user 453 ms, sys: 3.66 ms, total: 456 ms
Wall time: 455 ms
```
Without flint that is:
```python
In [1]: M = randMatrix(100)

In [2]: %time ok = M.inv_mod(163)
CPU times: user 7.28 s, sys: 7.54 ms, total: 7.29 s
Wall time: 7.28 s
```
The difference here is that flint makes `GF(p)` something like 10-20x faster.

Currently python-flint's implementation of matrices over GF(p) is not used but that would make this a lot faster still:
```python
In [1]: M = randMatrix(100)

In [2]: %time ok = M.inv_mod(163)
CPU times: user 441 ms, sys: 8.31 ms, total: 449 ms
Wall time: 448 ms

In [3]: fM = M.to_DM().to_dense().rep.rep

In [4]: type(fM)
Out[4]: flint.types.fmpz_mat.fmpz_mat

In [5]: import flint

In [6]: fMp = flint.nmod_mat(fM, 163)

In [7]: %time ok = fMp.inv()
CPU times: user 4.02 ms, sys: 0 ns, total: 4.02 ms
Wall time: 4.12 ms
```
The flint routine is 100x faster than DomainMatrix for this case and this is primarily because it is implemented in C.

Currently DomainMatrix will only use python-flint's matrix types for dense matrices over ZZ and QQ but not GF(p). In python-flint there are two versions of GF(p): nmod and fmpz_mod with one being for small moduli (something like `<2^32` bits) and `fmpz_mod` being for larger moduli. The corresponding Flint matrix types are not both exposed by python-flint: it has `nmod_mat` but not `fmpz_mod_mat`. It would be better to add `fmpz_mod_mat` to python-flint first and then to add support for using these for `GF(p)` in DomainMatrix.